fix: don't try to delete a file from a table

+++ udata_hydra/analysis/exports.py
@@ -43,7 +43,7 @@ async def _run_export_job(
                     f"{df_name.lower()}_size": output.filesize,
                 },
             )
-            if delete_input and config.REMOVE_GENERATED_FILES:
+            if delete_input and config.REMOVE_GENERATED_FILES and getattr(data_object, "file_name"):
                 data_object.path.unlink()
     except Exception as e:
         if data_object.resource_id:
